+++ src/ru/beykerykt/lightapi/server/nms/paperspigot/PaperSpigot_v1_8_R3.java
@@ -144,6 +144,6 @@ public class PaperSpigot_v1_8_R3 implements INMSHandler {
 			return 100;
 		double var2 = to.locX - from.locX;
 		double var4 = to.locZ - from.locZ;
-		return var2 * var2 + var4 * var4;
+		return (int) (var2 * var2 + var4 * var4);
 	}
 }
